Xml 2 Jaume
- Se ha modificado fileBuilder.java a fileXmlBuilder.java
- Se ha creado el método static xheckXmlExists para incluir en el Constructor de los Xml-DAO que verifica que el archivo XML existe.
- Se ha creado el método static subElementoIdList que nos permite crear listados de ID de instancias que se listen dentro del objeto.
