+++ src/main/resources/xml/Proyectos.xml
@@ -26,7 +26,7 @@
         <trabajadorList/>
         <financiadorList/>
         <accionList/>
-        <fechaInicio>2020-01-25</fechaInicio>
+        <fechaInicio>2019-12-05</fechaInicio>
         <fechaFin>2020-07-08</fechaFin>
         <lineaAccion>2</lineaAccion>
         <subLineaAccionList>
